BMS/requirements.txt
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
mysql-connector-python==8.0.26
Werkzeug==2.0.1
pytest==7.0.1
python-dotenv==0.19.0
Flask-WTF==0.15.1
Flask-Login==0.5.0